avoid MSVC warnings

+++ python/cif.cpp
@@ -327,8 +327,8 @@ void add_cif(py::module& cif) {
     }, py::keep_alive<0, 1>())
     .def("__repr__", [](const Table::Row& self) {
         std::string items;
-        for (size_t i = 0; i != self.size(); ++i)
-          items += " " + (self.has((int)i) ? self[i] : "None");
+        for (int i = 0; (size_t)i != self.size(); ++i)
+          items += " " + (self.has(i) ? self[i] : "None");
         return "<gemmi.cif.Table.Row:" + items + ">";
     });
